Add a local copy of the softfloat library
This library is not packaged on most platforms. This change adds a local copy
that is built by default. There is a flag, `UseBundledC`, that can be set to
`False` to use the previous behavior of expecting a system copy of the library.

Note that the local copy is only used for Linux and Mac builds right now. There
are configurations for Windows that should be possible with testing.

This change also enables github actions to make sure that the build works across
the supported platforms.