softfloat-hs.cabal: Sync c-sources with softfloat's Makefile
The C files listed in the `c-sources` were not in sync with what is actually
needed, as confirmed by cross-referencing the files with what is listed in
`softfloat`'s `Makefile`. The missing C dependencies in the `c-sources` were
the root cause of #9.

This fixes #9 in a minimal way by manually inspecting the differences between
the `c-sources` and the `Makefile` and fixing up the former to be in sync with
the latter. In a subsequent commit, I will make the connection between the
`c-sources` and the `Makefile` more obvious.