softfloat-hs.cabal: Structure C compiler options to more closely mirror softfloat Makefiles
This rearranges the C compiler–related options in the `.cabal` file to more
closely resemble the order in which file dependencies are listed in the
corresponding Makefiles on the `softfloat` side. This makes it much easier to
tell at a glance whether the two are in sync, and this will hopefully make
issues like #9 less likely to occur in the future.

This also adds support for the options in the Linux-ARM-VFPv2-GCC Makefile
(supporting ARM64 Linux and macOS) and the Win64-MinGW-w64 Makefile (supporting
x86-64 Windows). Therefore, this fixes #10. (There are other Makefiles that
`softfloat` uses (e.g., Linux-RISCV64-GCC), but as I do not have a way to test
these Makefiles, I have not added support for them in `softfloat-hs`.)